+++ src/main/java/com/fpetrola/z80/instructions/DAA.java
@@ -39,6 +39,6 @@ public class DAA<T extends WordNumber> extends ParameterizedUnaryAluInstruction<
   };
 
   public DAA(OpcodeReference target, Register<T> flag) {
-    super(target, flag, (tFlagRegister, reg_A) -> daaTableAluOperation.executeWithCarry2(reg_A, reg_A, tFlagRegister.read(), tFlagRegister));
+    super(target, flag, (tFlagRegister, reg_A) -> daaTableAluOperation.executeWithCarry2(reg_A, reg_A, tFlagRegister.read().intValue(), tFlagRegister));
   }
 }
